// src/app/interfaces/jugador.interface.ts
export interface Jugador {
  SportsDataID: string;
  Status: string;
  TeamID: number;
  Team:string ;
  Jersey: number;
  PositionCategory:string;
  Position: string;
  FirstName: string;
  LastName: string;
  BirthDate: string;
  BirthCity:string;
  BirthState: string;
  BirthCountry: string;
  GlobalTeamID:number;
  Height: number;
  Weight: number;
}