// src/app/interfaces/teamstats.interface.ts
export interface TeamStats {
  StatID: number;
  TeamID: number;
  SeasonType: number;
  Season: number;
  Name: string;
  Team: string;
  Wins: number;
  Losses: number;
  OpponentPosition: null | any;
  Possessions: number;
  GlobalTeamID: number;
  Updated: string;
  Games: number;
  FantasyPoints: number;
  Minutes: number;
  Seconds: number;
  FieldGoalsMade: number;
  FieldGoalsAttempted: number;
  FieldGoalsPercentage: number;
  EffectiveFieldGoalsPercentage: number;
  TwoPointersMade: number;
  TwoPointersAttempted: number;
  TwoPointersPercentage: number;
  ThreePointersMade: number;
  ThreePointersAttempted: number;
  ThreePointersPercentage: number;
  FreeThrowsMade: number;
  FreeThrowsAttempted: number;
  FreeThrowsPercentage: number;
  OffensiveRebounds: number;
  DefensiveRebounds: number;
  Rebounds: number;
  OffensiveReboundsPercentage: null | any;
  DefensiveReboundsPercentage: null | any;
  TotalReboundsPercentage: null | any;
  Assists: number;
  Steals: number;
  BlockedShots: number;
  Turnovers: number;
  PersonalFouls: number;
  Points: number;
  TrueShootingAttempts: number;
  TrueShootingPercentage: number;
  PlayerEfficiencyRating: null | any;
  AssistsPercentage: null | any;
  StealsPercentage: null | any;
  BlocksPercentage: null | any;
  TurnOversPercentage: null | any;
  UsageRatePercentage: null | any;
  FantasyPointsFanDuel: number;
  FantasyPointsDraftKings: number;
  FantasyPointsYahoo: number;
  PlusMinus: number;
  DoubleDoubles: number;
  TripleDoubles: number;
  FantasyPointsFantasyDraft: number;
  IsClosed: boolean;
  LineupConfirmed: null | any;
  LineupStatus: string;
  OpponentStat: OpponentStat | null;
}

export interface OpponentStat {
  StatID: number;
  TeamID: number;
  SeasonType: number;
  Season: number;
  Name: string;
  Team: string;
  Wins: number;
  Losses: number;
  OpponentPosition: null | any;
  Possessions: number;
  GlobalTeamID: number;
  Updated: string;
  Games: number;
  FantasyPoints: number;
  Minutes: number;
  Seconds: number;
  FieldGoalsMade: number;
  FieldGoalsAttempted: number;
  FieldGoalsPercentage: number;
  EffectiveFieldGoalsPercentage: number;
  TwoPointersMade: number;
  TwoPointersAttempted: number;
  TwoPointersPercentage: number;
  ThreePointersMade: number;
  ThreePointersAttempted: number;
  ThreePointersPercentage: number;
  FreeThrowsMade: number;
  FreeThrowsAttempted: number;
  FreeThrowsPercentage: number;
  OffensiveRebounds: number;
  DefensiveRebounds: number;
  Rebounds: number;
  OffensiveReboundsPercentage: null | any;
  DefensiveReboundsPercentage: null | any;
  TotalReboundsPercentage: null | any;
  Assists: number;
  Steals: number;
  BlockedShots: number;
  Turnovers: number;
  PersonalFouls: number;
  Points: number;
  TrueShootingAttempts: number;
  TrueShootingPercentage: number;
  PlayerEfficiencyRating: null | any;
  AssistsPercentage: null | any;
  StealsPercentage: null | any;
  BlocksPercentage: null | any;
  TurnOversPercentage: null | any;
  UsageRatePercentage: null | any;
  FantasyPointsFanDuel: number;
  FantasyPointsDraftKings: number;
  FantasyPointsYahoo: number;
  PlusMinus: number;
  DoubleDoubles: number;
  TripleDoubles: number;
  FantasyPointsFantasyDraft: number;
  IsClosed: boolean;
  LineupConfirmed: null | any;
  LineupStatus: string;
  OpponentStat: null;
}